Throw RuntimeException on UTF-16 conversion errors
Convert UnsupportedEncodingException into RuntimeException in interfaces
that expect Strings and unconditionally serialize them as UTF-16 bytes.
If the implementation does not support UTF-16 -- which is highly
unlikely, given that Java pioneered Unicode and UTF-16 specifically --
then the user application cannot do anything about it (other that
rewrite JavaThemis code). This is not an error with user input. Themis
does not give the user an option to select encoding and always uses
UTF-16, so the user application cannot recover from this error.