Wipe partial PRNG output
While this is only a theoretically possible attack vector, let's make
sure that we are not susceptible to it if it ever becomes practical.

If RAND_bytes() returns an error, wipe the output buffer clean so that
none of the partial output is available to the caller, thus avoiding
any possible PRNG state leaks *and* leaving highly conspicuous zeros
if the caller does not check for soter_rand() success.

For BoringSSL this is excessive because BoringSSL aborts the process
on PRNG failures. However, OpenSSL's built-in PRNG can make partial
writes into the buffer in case of failure.

Note that normally PRNG should not fail mid-flight. It's either
completely unusable because your system and/or OpenSSL is broken,
or it is able to produce arbitrary number of pseudorandom bytes.
There are special cases like very early boot stages on Linux, but
that behavior depends on whatever OpenSSL does in RAND_bytes(), and
that's... complicated. Just know that it *may* fail on some systems.