Warn about unused result of soter_rand()
Checking random number generation for failures is pretty imporant [1, 2]
so let's make it easier for users to track such issues. Introduce a
SOTER_MUST_USE macro that adds an attribute to the function which will
cause a compiler warning if the return value is not used (or not
explicitly ignored by casting to void). See [3] for description.

Since we treat warnings as errors, CI compilation will fail like this:

    compile build/obj/tests/common/test_utils.c.o
    tests/common/test_utils.c:70:3: warning: ignoring return value of function declared with 'warn_unused_result' attribute [-Wunused-result]
                    soter_rand((uint8_t*)(&res), sizeof(size_t));
                    ^~~~~~~~~~ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    1 warning generated.

if anyone adds a soter_rand() call without checking the return value.

[1]: https://github.com/veorq/cryptocoding#use-strong-randomness
[2]: http://jbp.io/2014/01/16/openssl-rand-api
[3]: http://releases.llvm.org/9.0.0/tools/clang/docs/AttributeReference.html#nodiscard-warn-unused-result