Add missing tests for themispp::secure_rand_t
We have this API, but it's not tested (or documented either). However,
it's there and we should at least exercise it in the test suite. Turns
out that it was not checking return value of soter_rand() and it's not
visible if there are no tests.

The API is... wacky, but that's what it is. I'm not really motivated to
clean it up at the moment (and that's out of scope).