Update PHP Composer checksum (#566)
Yes, they made a new release. We should probably revisit our pinning
approach with PHP Composer...

(cherry picked from commit a3abce44fc9fdcf71d5dfa0be11723dd0843e227)

This isn't really *needed* in this branch, but... it is because PHP
build needs this to be working, so we pull this and the next commit.