Fetch Composer SHA sum from GitHub
PHP Composer installer changes with every release and has a different
hash sum. Currently we hardwire the expected value in our installation
scripts. This leads to our CI builds breaking with every release of
Composer. While it's a nice way to get heads up about updates, it is
also somewhat stressful for the maintainers to be notified about that
with a broken build.

Instead of relying on a fixed hash sum, fetch it from GitHub Pages site
maintained by Composer. This is recommended way to install Composer
automatically [1].

This approach is arguably less secure because the attackers can
compromise GitHub Pages site to feed us the SHA value they need, but
given that Themis source code is also hosted on GitHub, a joint attack
on PHP infrastructure like that would probably be the least of our
concerns.

Moreover, Composer is used only when running PHPThemis unit tests. It is
not used during package building process, so it's currently not possible
to exploit the attack on Composer installer in order to, say, steal our
signing keys when building PHPThemis package.

Overall, I'd trade the possible risk for fewer randomly red builds and
not having to update a magical check sum every now and then.

[1]: https://getcomposer.org/doc/faqs/how-to-install-composer-programmatically.md