Use "fastcomp" flavor of Emscripten
We have figured out that it's not the new version that's broken, it's
the flavor. WasmThemis does not work well with the "upstream" flavor
of Emscripten which is the default for "lastest" branch since 1.39.

Instead of pinning an older version, use a suitable flavor. We'll be
testing with "upstream" later once we figure out what's wrong with it.