Fix Cargo with "--features"
No idea what changed here, but for some reason Cargo does not allow to
use "--features" with virtual crates now. However, it has no issues with
it once we're in "themis" crate subdirectory and ask to document all
crates in the workspace anyway.

Apparently, "--features" was not supposed to work with workspaces in the
first place [1], but I have no clue it worked before.

[1]: https://github.com/rust-lang/cargo/issues/5015