Reorder SOTER_MUST_USE to make Clang happy
Make sure that it goes first because when compiling for C++14 this gets
expanded into

    __attribute__((visibility("default")))
    [[nodiscard]]
    soter_status_t soter_rand(uint8_t* buffer, size_t length);

which throws Clang off-guard with this GCC compatibility syntax for
attributes. For some reason Clang thinks that [[nodiscard]] applies
only to "soter_status_t" which is not allowed.

Make sure that C++ attributes go before non-standard attributes.