Run ThemisPP tests with C++14 and C++17 (#572)
Explicitly test compatibility with those standard levels. C++20 is
currently being finalized. It's somewhat supported in the compilers,
but not in the versions shipped with Ubuntu 16.04.

It turned out that behavior wrt attributes might be different between
GCC and Clang, so let's test both compilers now.

Clang that we have in the Docker image does not support "-std=c++17",
so we do not test with that. (Actually, GCC support the option, but
__cplusplus value suggests that it's still in C++14 mode.)