Public PBKDF2 API in Soter
Add a new function to compute PBKDF2 with HMAC-SHA-256. We will use this
in the new passphrase mode of Secure Cell in Themis. It also might come
handy for other use cases so make this function public.

Mind the casts and integer ranges. OpenSSL uses "int" parameters,
BoringSSL uses "size_t" for sizes and "unsigned" for iteration count.
Soter consistently uses either size_t or explicitly-sized integers.
Check for possible overflows when casting. They should not be possible
with typical values, but just make sure we don't compute garbage.

Also note that this is not the same function as soter_pbkdf2() which can
be found in soter_sym.c files in OpenSSL/BoringSSL directories. That one
uses fixed iteration count, Secure Message uses it for key stretching.
And its definitely not the same as soter_kdf() which implements ZRTP KDF.