PBKDF2 test vectors
As said in the comment, there seems to be no RFC with test vectors for
PBKDF2 with HMAC-SHA2 specifically. There are test vectors for PBKDF2
with HMAC-SHA1, or for HMAC-SHA-256 on its own, but not in combination.
Therefore we use best available substitute.

Also throw is some API tests to verify that we reject invalid params.