Make "passphrase" required argument
It turns out that older OpenSSL versions (1.0.2) do not allow both
passphrase and salt to be empty. Modern versions as well as BoringSSL
allow this, but honestly it does not make much sense from practical
perspective.

Therefore, Soter also does not allow passphrase to be NULL or empty.
Salt argument is still optional though.