Tests for passphrase API
Here are some basic tests that are going to hold for the new API.

We verify a bunch of things:

- ability to encrypt and decrypt data
- reaction to invalid argument values
- incompatibility of passphrase and master-key API
- compatibility with older formats

We currently do *not* verify that Secure Cell is able to detect data
corruptions, invalid passphrases, etc. These tests will be added later.

Note the stability test suite. Some of language wrappers have something
like this but Themis Core generally did not verify that old data formats
are still supported. These tests verify that we are able to decrypt data
encrypted be previous versions of Themis. Passphrase API is going to
evolve and use different KDF configurations, effectively changing the
encryption algorithm. We need to be sure that we are still able to make
sense of the data after we change the implementation.

The tests also verify that we understand various AES key lengths that
may be used by different Themis builds. Existing Secure Cell code for
master key API handles them incorrectly. We'll fix it later.

And while we're here, move ARRAY_SIZE utility into common header file so
that it can be used across all tests.