Define authentication token format
Describe data structures used for Secure Cell authentication tokens.

Here we use much more explicit method than the existing master key
definitions. If we were to use the old approach, these structs would
look like

    struct themis_scell_auth_token_passphrase {
        uint32_t alg;
        uint32_t iv_length;
        uint32_t auth_tag_length;
        uint32_t message_length;
        uint32_t kdf_context_length;
    };
    // followed by IV data, auth tag data, KDF context

    struct themis_scell_pbkdf2_context {
        uint32_t iteration_count;
        uint16_t salt_length;
    };
    // followed by salt data

And then we'd go and cast byte buffers into this structs:

    (const struct themis_scell_seal_passphrase_header*)buffer
    buffer->iv_length;

While these structure definitions are definitely more compact, 'parsing'
code ends up scattered around the file and riddled with hard-to-reason
pointer magic. Furthermore, such pointer casts do not handle endianness
and pointer alignment correctly, leading to undefined behavior on some
ARM systems (and warnings from undefined behavior sanitizer -- that we
currently squelch because of such not-exactly-correct code).

New code is much more verbose -- in this file -- but the 'end-user' code
in Secure Cell implementation looks much cleaner and way easier to read
with all this complexity hidden elsewhere. The compiler does good job
at inlining all of the parsing code and it is as efficient as possible.
(Though yeah, it is *not* zero-copy as it was before.)

We put definitions in a private header file and make all parsing and
serialization functions "static inline" so that they do not cause
duplicate symbol errors and are always available to the compiler.

All of these are internal helpers and they are not intended to be reused
for other encryption modes or for any other purposes. They have rather
idiosyncratic API that might appear weird at first. Let it sink in.