Public PBKDF2 API (#574)
* Public PBKDF2 API in Soter

Add a new function to compute PBKDF2 with HMAC-SHA-256. We will use this
in the new passphrase mode of Secure Cell in Themis. It also might come
handy for other use cases so make this function public.

Mind the casts and integer ranges. OpenSSL uses "int" parameters,
BoringSSL uses "size_t" for sizes and "unsigned" for iteration count.
Soter consistently uses either size_t or explicitly-sized integers.
Check for possible overflows when casting. They should not be possible
with typical values, but just make sure we don't compute garbage.

Also note that this is not the same function as soter_pbkdf2() which can
be found in soter_sym.c files in OpenSSL/BoringSSL directories. That one
uses fixed iteration count, Secure Message uses it for key stretching.
And its definitely not the same as soter_kdf() which implements ZRTP KDF.

* PBKDF2 test vectors

As said in the comment, there seems to be no RFC with test vectors for
PBKDF2 with HMAC-SHA2 specifically. There are test vectors for PBKDF2
with HMAC-SHA1, or for HMAC-SHA-256 on its own, but not in combination.
Therefore we use best available substitute.

Also throw is some API tests to verify that we reject invalid params.

* Don't run long tests by default

16 million PBKDF2 iterations is a bit too long for casual test suite
runs. We run the test suite multiple times in "analyze" test job which
takes way too long to execute with this. Running then during development
all the time is also annoying.

Skip big iteration counts by default, but make sure that some CI jobs
run the full test suite.

* Add changelog entry

* Make "passphrase" required argument

It turns out that older OpenSSL versions (1.0.2) do not allow both
passphrase and salt to be empty. Modern versions as well as BoringSSL
allow this, but honestly it does not make much sense from practical
perspective.

Therefore, Soter also does not allow passphrase to be NULL or empty.
Salt argument is still optional though.

* Clarify key rotation issues

It is really good idea to update iteration count sometimes. However,
doing so results in a different derived key (obviously) and entails
all stuff related to key migration schemes. Make is clear in the
comments.