Validate key length on decryption
This is the first crash found by AFL. Decryption code fails to validate
key length encoded in "algorithm" field of Secure Cell header and just
proceeds to use it, assuming that it's one of 128, 192, 256 bits.

Maximum value allowed by data format is 4096 bits (512 bytes), but the
code allocates only 32 bytes for the derived key. Longer keys will cause
a buffer overflow which allows an attacker to overwrite return address
stored on the stack and to theoretically execute arbitrary code.
Practically, this can be used for denial of service by exploiting the
much more likely outcome -- crash due to a segmentation fault.

Fix the issue by validating the key length and allowing only values
produced by Themis: 128, 192, or 256 bits.