Improve AFL composability
Instead of using our special AFL_CFLAGS and AFL_LDFLAGS, just set them
per-target as we do in other places. We'd like to build AFL fuzzers
with sanitizers, and we'd like to be aware when AFL toolchain does not
support a particular compilation flag. Currently we can WITH_ASAN=1,
but it silently drops sanitizer flags instead of complaining, which
gives an impression that we use sanitizer when in fact we do not.