Exercise fuzzers during CI builds
Introduce a new "fuzz" job that builds, tests, and runs fuzzers.
We check that we can build all interesting configurations and run
the most interesting ones for a bit to ensure that there are no
obvious mistakes in the code. We'd like to run longer, but CircleCI
has around 20-minute limit which does not fit what we want to eat.

AFL is meant to be used interactively (primarily) but it is possible
to script its execution too. We use a bit peculiar spell for that:

    timeout -s INT 30s make fuzz | cat -u || true

with the following rationale:

- timeout limits execution time of "make fuzz" which is by default
  unbounded. Put a 30-second limit on execution to keep it civil.
  We've got quite a few fuzzing tools to run.

- AFL expects to be cleanly stopped by SIGINT. However, it will stop
  with non-zero exit code, so we eat it with "|| true"

- Piping into cat disconnects AFL from terminal so it won't be
  screaming about "terminal window is too small, plz resize it
  to at least 80x25, kthx"). Everything is better with cats.