Run benchmarks on CI
Do this for every build, at least to check that they are still
buildable. Also run some of them for a modest amount of time and attach
benchmark reports to builds.

It would be nice to compare base branch with PR branch to see if there
are any changes in that. However, with current approach branch switching
is kinda involved so we'll add that later.