Configurable PBKDF2 iteration count
KDF makes computations slow but fuzzing requires a lot of encryption and
decryption. Let's make the default iteration count to be configurable at
compile time and use a lower number for fuzzer builds. This drastically
increases efficiency of AFL search.

This feature is intended to be used mostly for debugging, and maybe to
provide custom builds on demand, so it's not documented anywhere except
for the code.