Changed needed to necessary, removed redundant words
Had a lengthy conversation with native English speaking tech writers about "needed vs necessary" recently. It's much better to use "necessary" in such situations as it is as a legit adjective and "needed" is perceived as a verb forced to act as an adjective in their eyes. That's not a set-in-stone rule, but "necessary" is just better here.