Automated Themis Core benchmarks (#580)
* Themis Core benchmarks

Let's add some reproducible, verifiable, and automatable benchmarks for
Themis performance. We are using Criterion.rs -- a framework in Rust --
but since C is easily callable from Rust, it's not much an issue to use
that framework for accurately benchmarking C code. And I don't have to
reinvent the wheel and build a benchmarking harness myself as it is
usually done in C/C++ world.

Currently we cover Secure Cell in Seal mode with master key. More
benches will come in later, this is only an example and a primer.
We exercise both encryption and decryption code paths. The data does not
really matter so we use whatever is easy to get. Message size comparison
confirms that encryption and decryption have roughly the same cost and
indeed have linear complexity.

* Run benchmarks on CI

Do this for every build, at least to check that they are still
buildable. Also run some of them for a modest amount of time and attach
benchmark reports to builds.

It would be nice to compare base branch with PR branch to see if there
are any changes in that. However, with current approach branch switching
is kinda involved so we'll add that later.