themispp::impl::input_buffer utility
This is how you say "AsRef<[u8]>" in C++. Yes, really.

We are going to use this bunch of templates to accept anything that
can be converted into a byte slice in new Secure Cell interface.

Public interface:

- themispp::input_buffer() templates

Application code is allowed to invoke these freely. They are resilient
to silly values, but until C++20 it is impossible to detect contiguous
iterators reliably. Therefore is is technically possible to pass, say,
std::deque<uint8_t> with inevitable probable segfault. Don't do that.
(Thankfully, std::list and everything non-random-access is ruled out.)

Private interface:

- themispp::impl::input_buffer struct
- themispp::impl::input_bytes() templates
- themispp::impl::input_string() templates

These functions are going to be used by ThemisPP internally. There are
dummy identity conversion to handle themispp::input_buffer() results.
String functions are used only by passphrase API constructors. They are
not available to general users so that they don't pass std::string
wherever they want.

However, it is possible for users to provide specializations of all
template functions in order to support their own custom containers.
Tests provide an example of how this can be done.

Note that the implementation is hidden in themispp::impl namespace
which itself is mirrored in "themispp/impl" directory. This should
prevent users from accidentally including and using these definitions.

Also note the "compat" headers which are used to polyfill newer features
of C++ when available. They are meant for internal use and should be
used in pairs to avoid leaking magic macros to application code.