themispp::secure_cell_context_imprint API
This one is easy to do: there are no overloads, simple API, there will
never be a passphrase API for Context Imprint. So it's mostly copy-paste
of the Seal mode. Just pay attention to the argument order in Core API.
It's slightly different for Context Imprint mode.

Since Context Imprint mode does not verify message validity, there is
not much that we can verify in tests. But at least we check the GIGO
behavior.