themispp::secure_cell_token_protect API
Ah! Token Protect API! The foster child of APIs in Themis which usually
does not get enough attention and ends up as a second-class citizen.
ThemisPP has specially obnoxious API for it, which is unique across all
Themis wrappers. Well, not anymore.

Instead of doing all that "set_token()"/"get_token()" we now return the
encryption results as std::pair (with a couple of better accessors
bolted onto it). This also gives us nice destructuring API which has
finally arrived in C++17.

Other than that, the implementation is pretty unremarkable. Mind the
argument order and you'll be fine.

The tests are templatized like Seal mode to make it easier to add
passphrase support later. Note that we very message and token
separately.