Mark old Secure Cell API deprecated
Yes, placement of attributes is... questionable but that's how you do it
in C++. Don't ask me, I have no clue and don't want to know. There is
probably a 75-page paper in C++ committee proceedings which explains in
great detail why it must be done this way and absolutely cannot be done
in a different, more sane way.

clang-format does not make it look pretty either. Oh well... At least
this code looks ugly and *definitely* deprecated.